+++ src/skyborn/plot/__init__.py
@@ -6,12 +6,18 @@ Created: 2026-03-01 14:58:56
 """
 
 from .contour import arrow_contour, arrow_contour_clabel, shadow_contourf
-from .plotting import add_equal_axes, createFigure, gradient_fill_between
+from .plotting import (
+    add_centered_axes,
+    add_equal_axes,
+    createFigure,
+    gradient_fill_between,
+)
 from .scatter import scatter
 from .vector import curly_vector, curly_vector_key
 
 __all__ = [
     "add_equal_axes",
+    "add_centered_axes",
     "arrow_contour",
     "arrow_contour_clabel",
     "createFigure",